requirements.txt
mistralai
python-dotenv
mcp
griffe>=1.7.3,<2.0
fastapi
uvicorn
httpx
